Group_4/election_predictor/requirements.txt
autocorrect==0.3.0
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-appconf==1.0.3
django-bootstrap-form==3.4
django-ipware==2.1.0
django-mysql==3.0.0.post1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.3
feedparser==5.2.1
geographiclib==1.49
idna==2.8
maxminddb==1.4.1
mysqlclient==1.4.2.post1
nltk==3.4.1
numpy==1.16.3
pandas==0.24.2
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
urllib3==1.24.2